feat: undo frame and modify enum
